refactor: move sdk class to right path
