+++ common-test/src/main/java/io/customer/commontest/BaseIntegrationTest.kt
@@ -1,7 +1,7 @@
 package io.customer.commontest
 
-import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOConfig
+import io.customer.sdk.android.CustomerIO
 import io.customer.sdk.data.model.Region
 import io.customer.sdk.extensions.random
