+++ core/src/main/kotlin/io/customer/sdk/CustomerIOInstance.kt
@@ -1,0 +1,9 @@
+package io.customer.sdk
+
+/**
+ * Allows mocking of CustomerIO for automated tests in the project.
+ * Mock the implementation of this interface to test the behavior of the SDK without actually calling the SDK.
+ */
+interface CustomerIOInstance {
+    val sdkVersion: String
+}
