+++ messaginginapp/src/main/java/io/customer/messaginginapp/ModuleMessagingInApp.kt
@@ -4,8 +4,8 @@ import android.app.Application
 import androidx.annotation.VisibleForTesting
 import io.customer.messaginginapp.di.gistProvider
 import io.customer.messaginginapp.hook.ModuleInAppHookProvider
-import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOConfig
+import io.customer.sdk.android.CustomerIO
 import io.customer.sdk.core.module.CustomerIOModule
 import io.customer.sdk.data.request.MetricEvent
 import io.customer.sdk.di.CustomerIOComponent
