+++ messagingpush/api/messagingpush.api
@@ -122,7 +122,7 @@ public final class io/customer/messagingpush/data/model/CustomerIOParsedPushPayl
 }
 
 public final class io/customer/messagingpush/di/DiGraphMessagingPushKt {
-	public static final fun pushMessaging (Lio/customer/sdk/CustomerIO;)Lio/customer/messagingpush/ModuleMessagingPushFCM;
+	public static final fun pushMessaging (Lio/customer/sdk/android/CustomerIO;)Lio/customer/messagingpush/ModuleMessagingPushFCM;
 }
 
 public final class io/customer/messagingpush/processor/PushMessageProcessor$Companion {
