+++ messagingpush/src/main/java/io/customer/messagingpush/ModuleMessagingPushFCM.kt
@@ -4,8 +4,8 @@ import androidx.annotation.VisibleForTesting
 import io.customer.messagingpush.di.fcmTokenProvider
 import io.customer.messagingpush.di.pushTrackingUtil
 import io.customer.messagingpush.lifecycle.MessagingPushLifecycleCallback
-import io.customer.sdk.CustomerIO
-import io.customer.sdk.CustomerIOInstance
+import io.customer.sdk.android.CustomerIO
+import io.customer.sdk.android.CustomerIOInstance
 import io.customer.sdk.core.module.CustomerIOModule
 import io.customer.sdk.di.CustomerIOComponent
