+++ samples/kotlin_compose/src/main/java/io/customer/android/sample/kotlin_compose/MainApplication.kt
@@ -8,7 +8,6 @@ import io.customer.android.sample.kotlin_compose.data.sdk.InAppMessageEventListe
 import io.customer.messaginginapp.MessagingInAppModuleConfig
 import io.customer.messaginginapp.ModuleMessagingInApp
 import io.customer.messagingpush.ModuleMessagingPushFCM
-import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOBuilder
 import javax.inject.Inject
 import kotlinx.coroutines.flow.first
@@ -29,7 +28,7 @@ class MainApplication : Application() {
         // TODO: Remove old builder and use new builder only to initialize the SDK
         // The new method should be called after the old method till the old method is removed
         // This is because the push and in-app modules are still using properties only initialized in the old method
-        CustomerIO.Builder(
+        io.customer.sdk.android.CustomerIO.Builder(
             siteId = configuration.siteId,
             apiKey = BuildConfig.API_KEY,
             appContext = this
