+++ samples/kotlin_compose/src/main/java/io/customer/android/sample/kotlin_compose/data/sdk/InAppMessageEventListener.kt
@@ -51,7 +51,7 @@ class InAppMessageEventListener(private val logger: Logger = Logger()) : InAppEv
         message: InAppMessage,
         arguments: Map<String, String>?
     ) {
-        io.customer.sdk.android.CustomerIO.instance().track(
+        CustomerIO.instance().track(
             "in-app message action",
             HashMap<String, String>().apply {
                 arguments?.let { putAll(it) }
