+++ samples/kotlin_compose/src/main/java/io/customer/android/sample/kotlin_compose/ui/customTrack/CustomAttribute.kt
@@ -98,7 +98,7 @@ fun CustomAttributeRoute(
                 Pair(
                     stringResource(R.string.send_profile_attribute)
                 ) {
-                    io.customer.sdk.android.CustomerIO.instance().profileAttributes = mapOf(attributeName to attributeValue)
+                    CustomerIO.instance().profileAttributes = mapOf(attributeName to attributeValue)
                 }
             } else {
                 Pair(stringResource(R.string.send_device_attribute)) {
